Небольшая оптимизация
